Add error triggers to authentication errors
The code for the ajax error handling function was modified in the authentication.js and authentication.min.js files. Specifically, jQuery's trigger method was added to emit an 'error' event whenever an authentication error occurs. This addition will make it more comfortable to handle such errors in the rest of the application.