Update error messages in index.php
The error messages displayed to the users have been updated in the index.php file to be more informative. This change will give users a clearer understanding of any issues, whether they involve missing login credentials or a failed request. The goal is to improve user experience and better guide user actions in case of an error.